fix(plugin): CodeRabbit review — CLAP param ABI, bounded drain, drop visibility (#238)
- CRITICAL: ClapEventParamValue::noteId was int16_t; clap_event_param_value_t
  ::note_id is int32_t. The wrong width misaligned `value`, so a real CLAP plugin
  would read a garbage parameter value (and past the struct). Fixed to int32_t.
  (The CI echo test bypasses the flush path, which is why this latent ABI bug in
  the real-CLAP path was not caught there.)
- drainParamQueueToChild now drains at most kMaxParamDrainPerPass (64) changes
  per worker pass; the remainder stays queued in order. A large preset burst
  against a slow child can no longer monopolize the worker and starve PROCESS
  (which would drop the RT callback to audible passThrough).
- Raised kParamQueueCapacity 512 -> 4096 to absorb full-preset bursts, added a
  parameterDropCount() getter, and log the first drop — setParameter is void, so
  queue overflow is now observable instead of silent.

SecOutOfProcessPluginHost still green.

+++ AestraAudio/src/Plugin/AestraPluginHostMain.cpp
@@ -150,7 +150,7 @@ struct ClapEventParamValue {
     ClapEventHeader header;
     uint32_t paramId;
     void* cookie;
-    int16_t noteId;
+    int32_t noteId; // clap_event_param_value_t::note_id is int32_t (not int16_t)
     int16_t portIndex;
     int16_t channel;
     int16_t key;
